update jlink windows support

+++ tockloader/jlinkexe.py
@@ -12,19 +12,14 @@ correct command line argument for future communication.
 '''
 
 import logging
+import os
+import platform
 import subprocess
 import tempfile
 import time
-import os
-import platform
 
 from .board_interface import BoardInterface
 from .exceptions import TockLoaderException
-
-jlink_exe_cmd = 'JLinkExe'
-
-if any(platform.win32_ver()):
-	jlink_exe_cmd = 'JLink'
 
 class JLinkExe(BoardInterface):
 	def _run_jtag_commands (self, commands, binary, write=True):
@@ -35,30 +30,45 @@ class JLinkExe(BoardInterface):
 		- `write`: Set to true if the command writes binaries to the board. Set
 		  to false if the command will read bits from the board.
 		'''
-		delete = True
+
+		# On Windows, the executable name is different.
+		executable_name = 'JLinkExe'
+		if platform.system() == 'Windows':
+			executable_name = 'JLink'
+
+		# On Windows, do not delete temp files because they delete too fast.
+		delete = platform.system() != 'Windows'
 		if self.args.debug:
 			delete = False
 
 		if binary or not write:
-			temp_bin = tempfile.NamedTemporaryFile(mode='w+b', suffix='.bin', delete=False)
+			temp_bin = tempfile.NamedTemporaryFile(mode='w+b', suffix='.bin', delete=delete)
 			if write:
 				temp_bin.write(binary)
 
-			temp_bin.close()
+			temp_bin.flush()
+
+			# On Windows we have set the files to not delete, so closing them
+			# will not cause them to be removed. However, we close them to avoid
+			# file locking issues.
+			if platform.system() == 'Windows':
+				temp_bin.close()
 
 			# Update all of the commands with the name of the binary file
 			for i,command in enumerate(commands):
 				commands[i] = command.format(binary=temp_bin.name)
 
-		jlink_file = tempfile.NamedTemporaryFile(mode='w', delete=False)
-		try:
+		with tempfile.NamedTemporaryFile(mode='w', delete=delete) as jlink_file:
 			for command in commands:
 				jlink_file.write(command + '\n')
 
-			jlink_file.close()
+			jlink_file.flush()
+
+			if platform.system() == 'Windows':
+				jlink_file.close()
 
 			jlink_command = '{} -device {} -if {} -speed {} -AutoConnect 1 -jtagconf -1,-1 -CommanderScript {}'.format(
-                                jlink_exe_cmd, self.jlink_device, self.jlink_if, self.jlink_speed, jlink_file.name)
+                                executable_name, self.jlink_device, self.jlink_if, self.jlink_speed, jlink_file.name)
 
 			if self.args.debug:
 				logging.debug('Running "{}".'.format(jlink_command))
@@ -86,14 +96,27 @@ class JLinkExe(BoardInterface):
 			if 'Error while programming flash' in stdout:
 				raise TockLoaderException('ERROR: Problem flashing.')
 
+		# On Windows we need to re-open files to do a possible read, and cleanup
+		# files that we could not set to auto delete.
+		if platform.system() == 'Windows':
+			ret = None
 			if write == False:
 				# Wanted to read binary, so lets pull that
 				with open(temp_bin.name, "rb") as temp_bin:
 					temp_bin.seek(0, 0)
-					return temp_bin.read()
-		finally:
-			os.remove(jlink_file.name)
-			os.remove(temp_bin.name)
+					ret = temp_bin.read()
+
+			# Cleanup files on Windows if needed.
+			if not self.args.debug:
+				os.remove(jlink_file.name)
+				os.remove(temp_bin.name)
+
+			return ret
+
+		if write == False:
+			# Wanted to read binary, so lets pull that
+			temp_bin.seek(0, 0)
+			return temp_bin.read()
 
 	def flash_binary (self, address, binary):
 		'''
@@ -215,9 +238,14 @@ class JLinkExe(BoardInterface):
 			logging.error('Unknown jlink_device. Use the --board or --jlink-device options.')
 			return
 
+		# On Windows, the executable name is different.
+		executable_name = 'JLinkExe'
+		if platform.system() == 'Windows':
+			executable_name = 'JLink'
+
 		logging.status('Starting JLinkExe JTAG connection.')
 		jtag_p = subprocess.Popen('{} -device {} -if {} -speed {} -autoconnect 1 -jtagconf -1,-1'.format(
-                    jlink_exe_cmd, self.jlink_device, self.jlink_if, self.jlink_speed).split(),
+                    executable_name, self.jlink_device, self.jlink_if, self.jlink_speed).split(),
 			stdout=subprocess.PIPE, stderr=subprocess.PIPE)
 
 		# Delay to give the JLinkExe JTAG connection time to start before running
